Split up large listeners
